:art: extract model types into their own files
