+++ src/app/(main)/about/team/_components/execteam/ExecCard.tsx
@@ -1,5 +1,5 @@
-import type { Exec } from "@/types/models";
 import placeholder from "@/assets/image/200x200.png";
+import type { Exec } from "@/types/models/Exec";
 //import ImageFit from "@/components/ImageFit";
 import ImageCover from "@/components/ImageCover";
 import ExecCardText from "./ExecCardText";
