+++ src/app/(main)/events/[slug]/_components/IndividualEvent.tsx
@@ -1,7 +1,7 @@
+import ImageFit from "@/components/ImageFit";
+import { Event } from "@/types/models/Event";
 import Image from "next/image";
 import Link from "next/link";
-import { Event } from "@/types/models";
-import ImageFit from "@/components/ImageFit";
 
 const IndividualEvent = ({ event }: { event: Event }) => {
     return (
