+++ src/app/(main)/events/_components/EventsSection.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import { useState } from "react";
-import { Event } from "@/types/models";
 import { EmptyListPlaceholder } from "@/components/EmptyListPlaceholder";
+import { Event } from "@/types/models/Event";
+import { useState } from "react";
 import { Category } from "../_data/events.data";
 import EventCard from "./EventCard";
 import EventCategoryFilter from "./EventCategoryFilter";
