+++ src/app/(main)/events/_data/events.data.ts
@@ -1,6 +1,6 @@
-import type { Event } from "@/types/models";
 import { getAllEvents } from "@/payload/collections/events/getAllEvents";
 import { parseEvent } from "@/payload/collections/events/parseEvent";
+import type { Event } from "@/types/models/Event";
 
 export type Category = {
     name: string;
