+++ src/app/(main)/events/_data/events2024/events2024.data.ts
@@ -1,4 +1,4 @@
-import { Event } from "@/types/models";
+import { Event } from "@/types/models/Event";
 import { wdccAGM2024 } from "./wdcc-agm-2024";
 import { wdccChiasma2024 } from "./wdcc-chiasma-2024";
