+++ src/app/(main)/events/_data/events2025/events2025.data.ts
@@ -1,4 +1,4 @@
-import { Event } from "@/types/models";
+import { Event } from "@/types/models/Event";
 import { wdccAGM2025 } from "./wdcc-agm-2025";
 
 // add all events to this array please!
