+++ src/app/(main)/projects/_components/ProjectCard.tsx
@@ -1,8 +1,7 @@
-import React from "react";
-import Link from "next/link";
-import { Project } from "@/types/models";
 import WDCCLogo from "@/assets/svg/WDCCLogo";
 import SuspenseImage from "@/components/SuspenseImage";
+import { Project } from "@/types/models/Project";
+import Link from "next/link";
 import { ProjectTag } from "./ProjectTag";
 
 type ProjectCardProps = {
