+++ src/app/(main)/projects/_components/ProjectsSection.tsx
@@ -1,7 +1,7 @@
 "use client";
 
+import { Project } from "@/types/models/Project";
 import { useState } from "react";
-import { Project } from "@/types/models";
 import { EmptyListPlaceholder } from "../../../../components/EmptyListPlaceholder";
 import { ProjectCard } from "./ProjectCard";
 import { ProjectFilter } from "./ProjectFilter";
