+++ src/payload/collections/events/EventsCollection.ts
@@ -1,6 +1,6 @@
-import type { CollectionConfig } from "payload";
-import { EventCategories } from "@/types/models";
+import { EventCategories } from "@/types/models/Event";
 import { SLUG } from "@/utils/enums/slug";
+import type { CollectionConfig } from "payload";
 
 export const EventsCollection: CollectionConfig = {
     slug: SLUG.EVENTS,
