+++ src/payload/collections/projects/ProjectsCollection.ts
@@ -1,7 +1,7 @@
-import type { CollectionConfig } from "payload";
-import { difficulties } from "@/types/models";
 import { technologiesList } from "@/assets/svg/technologies";
+import { difficulties } from "@/types/models/Project";
 import { SLUG } from "@/utils/enums/slug";
+import type { CollectionConfig } from "payload";
 
 export const ProjectsCollection: CollectionConfig = {
     slug: SLUG.PROJECTS,
