+++ src/payload/collections/projects/parseProject.ts
@@ -1,5 +1,5 @@
-import type { Project } from "@/types/models";
 import type { Project as CMSProject } from "@/payload/payload-types";
+import type { Project } from "@/types/models/Project";
 import { isNullish, removeDuplicates } from "@/utils/misc";
 import { media } from "@/utils/payload";
