+++ src/types/models/Faq.ts
@@ -1,0 +1,5 @@
+
+export type Faq = {
+    question: string;
+    answer: string;
+};
